Flatten file structure
