--- tests/hypervisor_comap/mod.rs
@@ -1,4 +1,0 @@
-// Every test file needs to be added here
-mod map_to_stack;
-mod stack_to_map;
-mod subsetting;
